restructure layout for embed, use new iframe links
